Finish a long journey to getting shifts to work
Currently lots of code in this commit and not at all tested. Not how I wanted to
do things but I wasn't sure how to get what I wanted to work. It has been tricky
to say the least. Been running in lots of circles trying to get the behavior
that I wanted. Finally have a shift created and associated with a schedule. Now
I just need to make it editable as well properly displaying the set schedule in
the view. My apologies to my inner self who is crying that this wasn't built
using TDD.