Make the creation of shifts more RESTful
This commit is across several files which is not ideal. I chose to do so
to keep it all in one commit due to the extent of the changes that I was
making. This commit moves the creation of shifts from the schedules
controller to the shifts controller. The original implementation worked
but was not at all RESTful.

In the schedules_controller, I removed the set_schedule action and moved
it to the index action in the shifts_controller. This action is acting
more like the new and edit actions rolled into one. When an admin clicks
the 'Set schedule' button, it hits the index action regardless if this
is to add shifts for the first time or to update created shifts. I can
take the pursuit of REST a step further by adjusting the button to hit
either the new or edit action depending on if the schedule already has
any created shifts or not.

The rest of the file changes were made to get the tests passing that
were broken when I made the above change.