Move add/remove employees to their own controller
The schedules_controller was handling the adding and removing of
employees being associated with the schedule. It only handled the
association between a schedule and employees, not creating or deleting
an employee.

I created a new controller called associate_employees to handle this. It
has create and destroy actions that handle the previous add_employees
and remove_employees from the schedules_controller. Updated the routes
and the url path in the views and all tests were passing.