Move employee status toggle to its own controller
The toggling of the employee status was previously in the
employees_controller under the toggle_status action. I created a new
controller called status_controller to handle this functionality. The
toggle_status was moved into the new controler under the update action.

All tests were green after updating the route and the url paths in the
views.