Change print_schedule to be a version of show
The print_schedule action was merely a modified schedule show page which
assigned a variable to hide a few buttons and modify the schedule layout
to hide the employees' availability. Instead of a completely separate
page, a changed the link for printing a schedule to assign a param
called `print: true`. When the show action would see the param, it would
assign the variable `@printable_view = true` which would control what
showed on the page.

I also created two helper methods to control what buttons would show on
the schedule show page. The `which_buttons?` method determines if the
page will show the "Add all employees" or "Remove all employees" button,
and the "Set schedule" button. The helper method `print_view?` will show
the "Print schedule" button unless the `@printable_view` variable has
been set.

The view and route changes reflect the above. I also removed the words
"Printable schedule" from the print page, and updated the test to
reflect this. The test now looks for the absence of the words "Set
schedule" as they should not show on the print schedule page.